Started working on a disk partition interface

+++ kernel/dev.c
@@ -115,7 +115,7 @@ int device_read(device_t* dev, uint64_t off, size_t len, void* buf)
     return dev->__dev_read(dev, off, len, buf);
 }
 
-int device_ioctl(device_t* dev, uint32_t op, uint32_t* ret, uint32_t data1, uint64_t data2)
+int device_ioctl(device_t* dev, uint32_t op, uint64_t* ret, uint32_t data1, uint64_t data2)
 {
     if (!dev) return 0;
     if (!dev->__dev_ioctl) {
